requirements.txt
arrow==0.12.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
blessed==1.15.0
certifi==2018.11.29
chardet==3.0.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1
django-compat==1.0.15
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.1.4
django-filter==2.0.0
django-haystack==2.8.1
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-picklefield==2.0
django-pytest==0.2.0
django-q==1.0.1
django-rest-swagger==2.1.0
django-social-share==1.3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-haystack==1.8.4
factory-boy==2.11.1
Faker==1.0.1
flake8==3.6.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
openapi==1.1.0
openapi-codec==1.3.2
packaging==18.0
pip-upgrader==1.4.6
pluggy==0.8.0
psutil==5.4.8
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.12.0
terminaltables==3.1.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
Whoosh==2.7.4
wrapt==1.10.11
